Extract schedule syncing logic into OrderCycleForm
